type is added to Online request

+++ database/migrations/2021_06_29_174804_create_pre_requests_table.php
@@ -16,7 +16,7 @@ class CreatePreRequestsTable extends Migration
         Schema::create('pre_requests', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('procedure_id');
-            $table->unsignedInteger('affair_id')->nullable();
+            $table->unsignedBigInteger('affair_id')->nullable();
             $table->string('name')->nullable();
             $table->longText('description')->nullable();
             $table->softDeletes();
